Initial Aura script conversion
