+++ Source/Game/Scripting/Interfaces/Aura/ICheckAreaTarget.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Game.Entities;
+
+namespace Game.Scripting.Interfaces.Aura
+{
+    public interface ICheckAreaTarget : IAuraScript
+    {
+        bool CheckAreaTarget(Unit target);
+    }
+}
